chore: Replace the old "Send" components with the new ones
