+++ packages/desktop/components/popups/Popup.svelte
@@ -41,8 +41,7 @@
     import RemoveNode from './RemoveNode.svelte'
     import RemoveProposalPopup from './RemoveProposalPopup.svelte'
     import RevotePopup from './RevotePopup.svelte'
-    // import { SendConfirmationPopup, SendFormPopup } from './send'
-    import { SendConfirmationPopup2, SendFormPopup2 } from './send'
+    import { SendConfirmationPopup, SendFormPopup } from './send'
     import StopVotingPopup from './StopVotingPopup.svelte'
     import BalanceBreakdownPopup from './BalanceBreakdownPopup.svelte'
     import TestDeepLinkFormPopup from './TestDeepLinkFormPopup.svelte'
@@ -128,10 +127,8 @@
         [PopupId.RemoveNode]: RemoveNode,
         [PopupId.RemoveProposal]: RemoveProposalPopup,
         [PopupId.Revote]: RevotePopup,
-        // [PopupId.SendConfirmation]: SendConfirmationPopup,
-        [PopupId.SendConfirmation]: SendConfirmationPopup2,
-        // [PopupId.SendForm]: SendFormPopup,
-        [PopupId.SendForm]: SendFormPopup2,
+        [PopupId.SendConfirmation]: SendConfirmationPopup,
+        [PopupId.SendForm]: SendFormPopup,
         [PopupId.StopVoting]: StopVotingPopup,
         [PopupId.BalanceBreakdown]: BalanceBreakdownPopup,
         [PopupId.TestDeepLinkForm]: TestDeepLinkFormPopup,
