+++ packages/desktop/components/popups/send/SendConfirmationPopup.svelte
@@ -1,165 +1,109 @@
 <script lang="ts">
     import { onMount } from 'svelte'
     import { get } from 'svelte/store'
+    import { FontWeight, Tab, TextHintVariant, TextType } from 'shared/components/enums'
     import {
+        ActivityInformation,
+        BasicActivityDetails,
         Button,
         ExpirationTimePicker,
         KeyValueBox,
+        NftActivityDetails,
         Text,
         TextHint,
         Toggle,
-        FontWeight,
-        TextType,
-        NftActivityDetails,
-        BasicActivityDetails,
-        ActivityInformation,
-    } from 'shared/components'
-    import { Tab, TextHintVariant } from 'shared/components/enums'
+    } from '@ui'
     import { prepareOutput, selectedAccount } from '@core/account'
     import { localize } from '@core/i18n'
     import { checkActiveProfileAuth, isActiveLedgerProfile } from '@core/profile'
     import { TimePeriod } from '@core/utils'
-    import { ActivityDirection, ActivityType, InclusionState, ActivityAction, TokenStandard } from '@core/wallet/enums'
-    import { newTransactionDetails, updateNewTransactionDetails, NewTransactionType } from '@core/wallet/stores'
+    import { TokenStandard } from '@core/wallet/enums'
+    import { newTransactionDetails, NewTransactionType, updateNewTransactionDetails } from '@core/wallet/stores'
     import { sendOutput } from '@core/wallet/actions'
     import {
-        getOutputParameters,
         validateSendConfirmation,
-        getAddressFromSubject,
+        getStorageDepositFromOutput,
+        getOutputParameters,
         getDefaultTransactionOptions,
     } from '@core/wallet/utils'
     import { closePopup, openPopup, PopupId, updatePopupProps } from '@auxiliary/popup'
-    import { Activity, NewTokenTransactionDetails } from '@core/wallet/types'
+    import { NewTokenTransactionDetails, NftActivity, TransactionActivity, VestingActivity } from '@core/wallet/types'
     import { CommonOutput, Output } from '@iota/sdk/out/types'
     import { ledgerPreparedOutput } from '@core/ledger'
-    import { getStorageDepositFromOutput } from '@core/wallet/utils/generateActivity/helper'
     import { handleError } from '@core/error/handlers/handleError'
-    import {
-        ACCOUNTS_CONTRACT,
-        CONTRACT_FUNCTIONS,
-        getDestinationNetworkFromAddress,
-        TARGET_CONTRACTS,
-        TRANSFER_ALLOWANCE,
-    } from '@core/layer-2'
     import { ToggleColor } from '@ui/inputs/Toggle.svelte'
+    import { getInitialExpirationDate, rebuildActivity } from '@core/wallet/utils/send/sendUtils'
 
     export let _onMount: (..._: any[]) => Promise<void> = async () => {}
+    export let isSendAndClosePopup: boolean = false
     export let disableBack = false
-    export let isCallbackFromUnlockStronghold: boolean = false
-    export let calculatedStorageDeposit: number = 0
+    export let preparedOutput: Output
 
-    let {
+    const transactionDetails = get(newTransactionDetails)
+    const {
+        type: transactionType,
         recipient,
-        expirationDate,
-        giftStorageDeposit,
-        surplus,
-        disableChangeExpiration,
-        disableToggleGift,
         layer2Parameters,
+        disableToggleGift,
+        disableChangeExpiration,
     } = get(newTransactionDetails)
 
-    let storageDeposit = calculatedStorageDeposit
-    let minimumStorageDeposit = 0
-    let preparedOutput: Output
+    let { surplus, expirationDate, giftStorageDeposit } = get(newTransactionDetails)
+
+    let activeTab: Tab
+    let activity: Partial<TransactionActivity | VestingActivity | NftActivity> | undefined = undefined
     let expirationTimePicker: ExpirationTimePicker
+    let initialExpirationDate: TimePeriod
+
+    let storageDeposit: number = 0
+    let minimumStorageDeposit = 0
     let visibleSurplus: number | undefined = undefined
 
-    let initialExpirationDate: TimePeriod
-    let activeTab: Tab
-
-    $: transactionDetails = get(newTransactionDetails)
-    $: isInternal = recipient.type === 'account'
     $: expirationTimePicker?.setNull(giftStorageDeposit)
+
     $: isBaseTokenTransfer =
         transactionDetails.type === NewTransactionType.TokenTransfer &&
         transactionDetails.asset?.metadata?.standard === TokenStandard.BaseToken
-    $: hideGiftToggle = isBaseTokenTransfer || !!layer2Parameters || (disableToggleGift && !giftStorageDeposit)
-    $: expirationDate, giftStorageDeposit, refreshSendConfirmationState()
+    $: isInternal = recipient.type === 'account'
     $: isTransferring = $selectedAccount.isTransferring
+    $: hideGiftToggle = isBaseTokenTransfer || !!layer2Parameters || (disableToggleGift && !giftStorageDeposit)
 
-    let activity: Activity | undefined = undefined
-    $: activity = {
-        ...transactionDetails,
-        assetId: transactionDetails.type === NewTransactionType.TokenTransfer ? transactionDetails.asset.id : undefined,
+    $: if (!isSendAndClosePopup) expirationDate, giftStorageDeposit, void rebuildTransactionOutput()
+
+    $: activity = rebuildActivity(
+        transactionDetails,
+        recipient,
         storageDeposit,
-        subject: recipient,
+        giftStorageDeposit,
+        visibleSurplus,
         isInternal,
-        id: undefined,
-        outputId: undefined,
-        transactionId: undefined,
-        time: undefined,
-        containsValue: true,
-        isAssetHidden: false,
-        asyncData: undefined,
-        giftedStorageDeposit: 0,
-        surplus: visibleSurplus,
-        type: ActivityType.Basic,
-        direction: ActivityDirection.Outgoing,
-        inclusionState: InclusionState.Pending,
-        action: ActivityAction.Send,
-        destinationNetwork: getDestinationNetworkFromAddress(layer2Parameters?.networkAddress),
-        ...(layer2Parameters?.networkAddress && {
-            parsedLayer2Metadata: {
-                ethereumAddress: getAddressFromSubject(recipient),
-                targetContract: TARGET_CONTRACTS[ACCOUNTS_CONTRACT],
-                contractFunction: CONTRACT_FUNCTIONS[TRANSFER_ALLOWANCE],
-                gasBudget: layer2Parameters?.gasBudget ?? 0,
-            },
-        }),
-    }
+        layer2Parameters
+    )
 
-    function refreshSendConfirmationState(): void {
-        if (!isCallbackFromUnlockStronghold) {
-            if (
-                transactionDetails.type === NewTransactionType.NftTransfer &&
-                Number($selectedAccount.balances.baseCoin.available) === 0
-            ) {
-                giftStorageDeposit = true
-                disableChangeExpiration = true
-                disableToggleGift = true
-            }
-            updateNewTransactionDetails({ type: transactionDetails.type, expirationDate, giftStorageDeposit, surplus })
-            void prepareTransactionOutput()
-        } else {
-            initialExpirationDate = getInitialExpirationDate()
+    onMount(async () => {
+        await updateStorageDeposit()
+        initialExpirationDate = getInitialExpirationDate(expirationDate, storageDeposit, giftStorageDeposit)
+
+        try {
+            await _onMount()
+        } catch (err) {
+            handleError(err)
         }
-    }
+    })
 
-    function getInitialExpirationDate(): TimePeriod {
-        if (expirationDate) {
-            return TimePeriod.Custom
-        } else if (storageDeposit && !giftStorageDeposit) {
-            return TimePeriod.OneDay
-        } else {
-            return TimePeriod.None
-        }
-    }
+    async function rebuildTransactionOutput(): Promise<void> {
+        updateNewTransactionDetails({
+            type: transactionType,
+            expirationDate,
+            giftStorageDeposit,
+        })
 
-    async function prepareTransactionOutput(): Promise<void> {
         try {
             const transactionDetails = get(newTransactionDetails)
             const outputParams = await getOutputParameters(transactionDetails)
             preparedOutput = await prepareOutput($selectedAccount.index, outputParams, getDefaultTransactionOptions())
 
             await updateStorageDeposit()
-
-            // Note: we need to adjust the surplus
-            // so we make sure that the surplus is always added on top of the minimum storage deposit
-            if (Number(surplus) > 0) {
-                if (minimumStorageDeposit >= Number(surplus)) {
-                    visibleSurplus = surplus = undefined
-                } else {
-                    visibleSurplus = Number(surplus) - minimumStorageDeposit
-                    // Note: we have to hide it because currently, in the sdk,
-                    // the storage deposit return strategy is only looked at
-                    // if the provided amount is < the minimum required storage deposit
-                    hideGiftToggle = true
-                }
-            }
-
-            if (transactionDetails.expirationDate === undefined) {
-                initialExpirationDate = getInitialExpirationDate()
-            }
         } catch (err) {
             handleError(err)
         }
@@ -168,22 +112,38 @@
     async function updateStorageDeposit(): Promise<void> {
         const { storageDeposit: _storageDeposit, giftedStorageDeposit: _giftedStorageDeposit } =
             await getStorageDepositFromOutput($selectedAccount, preparedOutput as CommonOutput)
+
         storageDeposit = minimumStorageDeposit = _storageDeposit > 0 ? _storageDeposit : _giftedStorageDeposit
+
         if (isBaseTokenTransfer) {
             const rawAmount = Number((transactionDetails as NewTokenTransactionDetails).rawAmount)
             if (rawAmount >= storageDeposit) {
                 storageDeposit = 0
             }
         }
+
+        // Note: we need to adjust the surplus
+        // so we make sure that the surplus is always added on top of the minimum storage deposit
+        if (Number(surplus) > 0) {
+            if (minimumStorageDeposit >= Number(surplus)) {
+                visibleSurplus = surplus = undefined
+            } else {
+                visibleSurplus = Number(surplus) - minimumStorageDeposit
+                // Note: we have to hide it because currently, in the sdk,
+                // the storage deposit return strategy is only looked at
+                // if the provided amount is < the minimum required storage deposit
+                hideGiftToggle = true
+            }
+        }
+    }
+
+    function toggleGiftStorageDeposit(): void {
+        giftStorageDeposit = !giftStorageDeposit
     }
 
     async function sendOutputAndClosePopup(): Promise<void> {
         await sendOutput(preparedOutput)
         closePopup()
-    }
-
-    function toggleGiftStorageDeposit(): void {
-        giftStorageDeposit = !giftStorageDeposit
     }
 
     async function onConfirmClick(): Promise<void> {
@@ -194,7 +154,7 @@
                 ledgerPreparedOutput.set(preparedOutput)
             }
 
-            updatePopupProps({ isCallbackFromUnlockStronghold: true, calculatedStorageDeposit: storageDeposit })
+            updatePopupProps({ isSendAndClosePopup: true, preparedOutput })
             await checkActiveProfileAuth(sendOutputAndClosePopup, { stronghold: true, ledger: false })
         } catch (err) {
             handleError(err)
@@ -212,14 +172,6 @@
     function onCancelClick(): void {
         closePopup()
     }
-
-    onMount(async () => {
-        try {
-            await _onMount()
-        } catch (err) {
-            handleError(err)
-        }
-    })
 </script>
 
 <send-confirmation-popup class="w-full h-full space-y-6 flex flex-auto flex-col shrink-0">
