+++ packages/desktop/components/popups/send/index.js
@@ -1,4 +1,2 @@
 export { default as SendFormPopup } from './SendFormPopup.svelte'
-export { default as SendFormPopup2 } from './SendFormPopup2.svelte'
 export { default as SendConfirmationPopup } from './SendConfirmationPopup.svelte'
-export { default as SendConfirmationPopup2 } from './SendConfirmationPopup2.svelte'
